+++ RELEASE_NOTES.md
@@ -2,8 +2,11 @@
 
 ## 2.7.2 2021-11-18
 
-* Increases number of retries when downloading manifests
 * Uses manifest files for large subjob inputs
+* Uses dxda to bulk-download manifest files
+* Increases number of retries when downloading single manifest files
+* Fixes error when parsing a field name with multiple `stage-*` prefixes
+* Allows file-to-string coercion for WDL inputs
 
 ### Dependency updates